calculator with 4 basic functions
calculator with 4 basic functions: addition, subtraction, multiplication and division (no implementation of negative numbers). Text recognition is implemented through recursive descent, the calculation occurs immediately, that is, no trees are matched, in case of a text recognition error, a syntax error occurs, a function for skipping spaces is also implemented, skipping spaces occurs an infinite number of times, a caveat, there is no function for moving to a new line.